Refactory recovery logic
